+++ src/main/java/com/project/instagramclonebackend/Repository/StatusRepo.java
@@ -3,6 +3,6 @@ package com.project.instagramclonebackend.Repository;
 import com.project.instagramclonebackend.Entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepo extends JpaRepository <Status, Integer> {
+public interface StatusRepo extends JpaRepository <Status, Long> {
 
 }
